Dont try to close gocode when its not found #1966

+++ src/goInstallTools.ts
@@ -355,13 +355,16 @@ export function installTools(missing: string[]) {
 			let closeToolPromise = Promise.resolve(true);
 			if (tool === 'gocode' || tool === 'gocode-gomod') {
 				closeToolPromise = new Promise<boolean>((innerResolve) => {
-					cp.execFile(getBinPath(tool), ['close'], {}, (err, stdout, stderr) => {
-						if (stderr && stderr.indexOf('rpc: can\'t find service Server.') > -1) {
-							outputChannel.appendLine('Installing gocode aborted as existing process cannot be closed. Please kill the running process for gocode and try again.');
-							return innerResolve(false);
-						}
-						innerResolve(true);
-					});
+					const gocodeToolPath = getBinPath(tool);
+					if (path.isAbsolute(gocodeToolPath)) {
+						cp.execFile(gocodeToolPath, ['close'], {}, (err, stdout, stderr) => {
+							if (stderr && stderr.indexOf('rpc: can\'t find service Server.') > -1) {
+								outputChannel.appendLine('Installing gocode aborted as existing process cannot be closed. Please kill the running process for gocode and try again.');
+								return innerResolve(false);
+							}
+							innerResolve(true);
+						});
+					}
 				});
 			}
